Safari-609.1.11.1/LayoutTests/platform/wk2/http/tests/security/drag-drop-local-file-expected.txt

CONSOLE MESSAGE: line 13: Not allowed to drag local resource: foobar
ALERT: PASS
Dragme 
